requirements.txt
Flask==2.2.5
requests
python-dotenv
transformers
torch